% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_secret_email}
\alias{get_secret_email}
\title{Get email associated with secret in local Renvironment for reuse}
\usage{
get_secret_email()
}
\description{
Get email associated with secret in local Renvironment for reuse
}
