% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_endpoints.R
\name{list_endpoints}
\alias{list_endpoints}
\title{List data end points for API}
\usage{
list_endpoints()
}
\value{
data frame of available endpoints
}
\description{
\code{list_endpoints} generates a convenient data.frame listing the endpoint name under "endpoint"
along with a high-level general description of the endpoint data under "description"
}
\details{
\code{list_endpoints} will list the available endpoints that return data from the Tidy Depot API along with
a description of the type of data provided.  It is a useful function to ensure that the name provided
as part of the "path" argument to \code{tidy_depo_api} function call.  If the name provided to the path
argument is incorrent then an error will be thrown and no API call returned.
}
\author{
Gabriel Rosenfeld, \email{gabriel.rosenfeld@nih.gov}
}
