% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{store_secret_credentials}
\alias{store_secret_credentials}
\title{Store secret credentials in local Renvironment for reuse}
\usage{
store_secret_credentials()
}
\description{
Store secret credentials in local Renvironment for reuse
}
