% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{tidy_depot_api}
\alias{tidy_depot_api}
\title{Generate a DEPOT tidy API request}
\usage{
tidy_depot_api(path, token, cohortId = "")
}
\arguments{
\item{path}{The name of an endpoint of interest (e.g. Biochemistry)}

\item{token}{The token string returned from get_token() function}

\item{cohortId}{The int string containing the cohortId of interest from DEPOT}
}
\value{
An object of class tidy_depot_api containing API call results
}
\description{
\code{tidy_depot_api} executes an api request for the path endpoint (e.g. Patient-Case) and
returns a tidy_depot_api class object with the API call results unless there is an error
}
\details{
This function depends upon setting up your user credentials to access the Tidy Depot API locally including your email address
and secret key using \code{store_secret_credentials} in Rstudio.  This will prompt you to store that information in a local .Renviron
file.  Afterwards, retrieve a unique token string using the \code{get_token} function which is passed as the token argument to the
\code{tidy_depot_api} function along with the path end point of interest.  Any errors will be returned instead of the API call notifying
that there may be an issue with either your credentials or the endpoint name.
}
\author{
Gabriel Rosenfeld, \email{gabriel.rosenfeld@nih.gov}
}
