% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{get_token}
\alias{get_token}
\title{Generate a token from the TB DEPOT tidy API}
\usage{
get_token(email_address = get_secret_email(), secret = get_secret())
}
\arguments{
\item{email_address}{email address associated with secret}

\item{secret}{secret used for generating API token}
}
\value{
The token string used for Authentication header in API calls
}
\description{
Generate a token from the TB DEPOT tidy API
}
