% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_request.R
\name{tidy_depot_api}
\alias{tidy_depot_api}
\title{Generate a DEPOT tidy API request}
\usage{
tidy_depot_api(path, token)
}
\arguments{
\item{path}{The name of an endpoint of interest (e.g. Biochemistry)}

\item{token}{The token string returned from get_token() function}
}
\value{
An object of class tidy_depot_api containing API call results
}
\description{
Generate a DEPOT tidy API request
}
